Mostly comment and docstring fixes.
Add macro with-protocol-style which is a safer version of set-protocol-style.
(But #'broadcast is still probably more modern and better).

Add &optional arg of node type to #'make-nodes.

Add a bit of sanity checking in #'visualize-nodes because
one of our examples was using it wrong and causing the SVG
output to not show edges.