Decrease lowlevel-application-handler and application-handler noise in log messages because we're never using them.
Decrease noise of :NO-TIMEOUT-HANDLER! because it's irrelevant
when we're not using replies.

#'make-gossip-actor fn was printing "Gossip Actor" redundantly. Fixed.

#'debug-level --> #'show-debug-p because the string "debug level" was
so common in the code base as to be confusing.

Rewrite #'show-debug-p (nee #'debug-level) completely so it's obvious
how it works, and it now enables *debug-level* to be set to t
to show all log messages. I'm astonished it wasn't this way all along.

#'write-as-string --> #'write-log-entry-as-string

Break out #'%edebug as its own function that can be called by callers who
themselves need to call #'show-debug-p in cases where e.g. it's
expensive to compute arguments for #'edebug and we don't want to do it
unless the debug-level requires it.

Add *friendly-logs* flag which if true will enable the text version of the log to show
relative microsecond timestamps which are much easier to read than absolute timestamps.

If *debug-level* is 5 or more, #'send-msg log entries now show
the real source and destination nodes (using their print-object methods
which presumably themselves use their UIDS) rather than merely their UIDs. Duh.